Bug 2031187. Update libjpegturbo to 3.1.0. r=gfx-reviewers,lsalzman
The result of running

./mach vendor --patch-mode none media/libjpeg/moz.yaml
./mach vendor --patch-mode only media/libjpeg/moz.yaml

Differential Revision: https://phabricator.services.mozilla.com/D294049

UltraBlame original commit: 5747f70d7651429839f582eea569599006fe2775

+++ media/libjpeg/simd/powerpc/jcsample-altivec.c
@@ -50,7 +50,7 @@ void jsimd_h2v1_downsample_altivec(JDIMENSION image_width,
   expand_right_edge(input_data, max_v_samp_factor, image_width,
                     output_cols * 2);
 
-  for (outrow = 0; outrow < v_samp_factor; outrow++) {
+  for (outrow = 0; outrow < (int)v_samp_factor; outrow++) {
     outptr = output_data[outrow];
     inptr = input_data[outrow];
 
@@ -107,7 +107,7 @@ jsimd_h2v2_downsample_altivec(JDIMENSION image_width, int max_v_samp_factor,
   expand_right_edge(input_data, max_v_samp_factor, image_width,
                     output_cols * 2);
 
-  for (inrow = 0, outrow = 0; outrow < v_samp_factor;
+  for (inrow = 0, outrow = 0; outrow < (int)v_samp_factor;
        inrow += 2, outrow++) {
 
     inptr0 = input_data[inrow];
